move responsibility for setting up trigger from trigger itself back to worker.
The code is still in trigger.c, but it's called directly by the worker to set up
a triggerconfig_t structure. this may seem backwards, but it allows us to have a
sqlite3 db for each worker: sqlite3 is emphatically not threadsafe, so we don't
want to be making calls from a child thread.

Clearly I'm missing Sean, this migh tbe the longest commit message I've ever written.