Fix race condition in creation of line listeners
If line listeners are created before the line publishers, 0MQ falls in a heap.
(Usually this would be ok - a tcp endpoint always exists, in some platonic sense,
even if it isn't open right now. An inproc endpoint can be checked, however.

The solution to this is to delay the main process until the line dispatcher
process has gone through one loop, thereby creating all the publishers it
needs to.